CHORE Update build task
Since I added webpack to the build process, it's a bit messy, especially when I
generate the xpi file for Mozilla. This branch will take care of this, so that it's
as easy as possible for me to develop and build and eventually build the XPI.

Generate CSS files from SCSS, moving everything build to the /build directory. Moved
the assets around as well at the same time (might need some tests to check if every
path is still up to date)

REF: [#84](https://trello.com/c/vz0VG5Vk/84)
REF: [#86](https://trello.com/c/qZVht6zj/86)

Details
-------
* CHORE Updated webpack to v2, and updated the config accordingly
* CHORE Now generate CSS from SCSS source files
* CHORE Moved popup/list.* files to popup/popup.*

Others
------
* CHORE Updated README for developers