FIX disconnect from pocket
The "disconnect from pocket" link in the settings was rarely working (but still
working from time to time). Was due to the fact that it's triggered when clicking on
a link, and it seems the default action of the link (redirect to Addons page) was
occuring before the lookup in local storage, and was breaking that.

REF: [#172](https://trello.com/c/5VHLVBwT)