FIX Contextual menu and click on link
If a page is already in the pocket list, then contextual "add to pocket" was
disabled. But it was also disabled if you tried to right-click on a link of the
page to add it to pocket, which was not expected. It's hard to dynamically change
the state of an entry (would need click event handler on all pages), so I keep this
entry always enabled.

REF: [#196](https://trello.com/c/MI5h4Ty1)