FEATURE Red badge on errors
So far, errors occuring on API calls where displayed in the popup with a red flash at
the bottom, but no indication was given to the user in case the popup was closed at
the moment of the error. We now flash the page action badge with a red "!" to let the
user know something went wrong.

REF: [#187](https://trello.com/c/wuk5r3ib)

Details
-------
* FEATURE: Add badge flash on item add/remove
* FEATURE: Add badge flash on retrieve items first/diff

Others
------
* REFACTOR: ES6 method format + shorter log msg in background.js