PERFS Popup building performances
When building the popup UI, if a retrieveDiff call was necessary, the popup UI was
built only upon reception of the diff, which resulted in a delay before display the
items list. Plus if pocket service was unreachable, you couldn't view the list at
all, which was not good.

REF: [#213](https://trello.com/c/qS5zAct9)

Details
-------
* FEATURE: Implemented the "update list" behaviour, seems to be working quite well.
  List is displayed immediately based on localStorage, and updated when a
  retrieveDiff is issued (updated = only change the proper parts of the DOM, don't
  rebuild everything)
* REFACTOR: Reorganized popup code, try to namespace most of the code into its module
  (pagination, top actions, top filter)