[tags] Fix multi-line tags input + fix bug with border disappearing when editing first item
Remove the fixed .item height. It was used mainly to allow CSS transitionning on
height (which does not work with auto value). I found out that you can use max-height
for that purpose. Here I set max-height to 6rem which is a bit more than the height
of an item when you're editing tags on two rows (which should be the maximum?)