refacto: moved stuff concerning the items list from PopupUI to PopupItemList.
This result in a PopupUI module responsible for calling all other UI modules, and a
clearer responsibility for PopupItemList that now handles all things related to...
yup, the items list.

Drawback: PopupItemList is now a huge module (600+ LOC) and should be broken down at
some point. But using a split that makes some sense. Which was not the case so far.