fix: disabled Settings caching which caused some bugs
Initially, the fix was related to a bug when "display count badge" was
disabled. I discovered that it was related to the Settings caching
inside a background script: if the settings were updated in the
meantime, the background scripts kept its outdated version of the
settings, which could lead to some bugs, such as showing the badge count
even though the settings was disabled. I also suspect it was the cause
of the automations not working at the first install: settings were
loaded with automations disabled, then user enabled them, but the bg
script kept the outdated disabled version.

Disabled the caching altogether: this code is accessing localStorage, I
doubt this will cause any performance issues.

Other:

- deps: updated webextension-polyfill