fix: try to make `addItem` more solid, since I'm receiving sentries
The sentry are caused by an error when using `add` (to append one single
item to Pocket). The error case is not documented, but I _suppose_ that
in case of an error, `response.item` is missing, which would make my
former code store `response.action_results` in the addedItems.
`response.action_results` does not exist when using the `add` endpoint,
which would make it an undefined value. Later on I try to map on this,
so that could be the reason for the sentries I'm seeing.